+++ lib/canonical_local_syssolaris.h
@@ -1,2 +1,2 @@
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str38,
-  (int)(long)&((struct stringpool_t *)0)->stringpool_str515,
+  (int)(ptrdiff_t)&((struct stringpool_t *)0)->stringpool_str38,
+  (int)(ptrdiff_t)&((struct stringpool_t *)0)->stringpool_str515,
