+++ lib/iconv_open1.h
@@ -81,7 +81,7 @@
         goto invalid;
       continue;
     }
-    ap = aliases_lookup(buf,bp-buf);
+    ap = aliases_lookup(buf, (unsigned int)(bp-buf));
     if (ap == NULL) {
       ap = aliases2_lookup(buf);
       if (ap == NULL)
@@ -171,7 +171,7 @@
         goto invalid;
       continue;
     }
-    ap = aliases_lookup(buf,bp-buf);
+    ap = aliases_lookup(buf, (unsigned int)(bp-buf));
     if (ap == NULL) {
       ap = aliases2_lookup(buf);
       if (ap == NULL)
